+++ docker-compose.yml
@@ -68,7 +68,7 @@ services:
      - "8983:8983"
     environment:
       - SOLR_OPTS=-XX:-UseLargePages
-      - SOLR_HEAP=800m
+      - SOLR_HEAP=3g
       - ZK_HOST=zoo1:2181,zoo2:2181,zoo3:2181
       - JAEGER_SERVICE_NAME=solr1  # not working ;-(
       - JAEGER_SAMPLER_TYPE=const
@@ -90,7 +90,7 @@ services:
      - "8984:8983"
     environment:
       - SOLR_OPTS=-XX:-UseLargePages
-      - SOLR_HEAP=800m
+      - SOLR_HEAP=3g
       - ZK_HOST=zoo1:2181,zoo2:2181,zoo3:2181
       - JAEGER_SERVICE_NAME=solr2
       - JAEGER_SAMPLER_TYPE=const
@@ -112,7 +112,7 @@ services:
      - "8985:8983"
     environment:
       - SOLR_OPTS=-XX:-UseLargePages
-      - SOLR_HEAP=800m
+      - SOLR_HEAP=3g
       - ZK_HOST=zoo1:2181,zoo2:2181,zoo3:2181
       - JAEGER_SERVICE_NAME=solr3
       - JAEGER_SAMPLER_TYPE=const
