+++ embeddings/app/main.py
@@ -1,8 +1,10 @@
 from fastapi import FastAPI
-from strans import router as router_strans
+from clip import router as router_clip
+from minilm import router as router_minilm
 
 app = FastAPI()
 
-app.include_router(router_strans.router, prefix="/strans")
+app.include_router(router_clip.router, prefix="/clip")
+app.include_router(router_minilm.router, prefix="/minilm")
