+++ solr/configsets/ecommerce/conf/schema.xml
@@ -17,8 +17,8 @@ for demo purposes
   <field name="img_500x500" type="string" indexed="false" stored="true"/>
   <field name="img_thumb" type="string" indexed="false" stored="true"/>
   <field name="date_released" type="date" indexed="true" stored="true" />
-  <field name="product_vector" type="knn_vector" indexed="true" stored="true"/>
-  <field name="product_image_vector" type="knn_vector" indexed="true" stored="true"/>
+  <field name="product_vector" type="knn_vector_384" indexed="true" stored="true"/>
+  <field name="product_image_vector" type="knn_vector_768" indexed="true" stored="true"/>
 
   <!-- we don't stem names -->
   <field name="supplier" type="text_general" indexed="true" stored="true"/>
@@ -94,7 +94,8 @@ for demo purposes
 
   <fieldType name="random" class="solr.RandomSortField" indexed="true"/>
 
-  <fieldType name="knn_vector" class="solr.DenseVectorField" vectorDimension="768" similarityFunction="cosine" knnAlgorithm="hnsw"/>
+  <fieldType name="knn_vector_768" class="solr.DenseVectorField" vectorDimension="768" similarityFunction="cosine" knnAlgorithm="hnsw"/>
+  <fieldType name="knn_vector_384" class="solr.DenseVectorField" vectorDimension="384" similarityFunction="cosine" knnAlgorithm="hnsw"/>
 
   <fieldType name="text_stemmed_en" class="solr.TextField" autoGeneratePhraseQueries="false"
              enableGraphQueries="false" positionIncrementGap="100" multiValued="false">
